Remove dead tests
